% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTopic.R
\name{CellTopic}
\alias{CellTopic}
\title{CellTopic}
\usage{
CellTopic(
  spot_celltype,
  spot_clusters,
  cluster = NULL,
  num_topics = NULL,
  percent = NULL,
  Binarization = FALSE,
  meta.cell = FALSE,
  k = NULL,
  method = "complete"
)
}
\arguments{
\item{spot_celltype}{A data frame of the result of deconvolution, row
is spot, col is celltype.}

\item{spot_clusters}{A data frame contains clustering information for spots,
row is spots.}

\item{cluster}{A character. Use the first column in spot_clusters as the
column name if \code{NULL}. Or provide your own column names that represent
clustering information in spot_clusters.}

\item{num_topics}{A integer or a vector of integer. The number of topics.
default is \code{NULL}, find the best number of topics by default. if a vector of
integer, find the best number of topics by the given range.}

\item{percent}{A numeric from \code{0} to \code{1}. The percent of topics. Default is \code{0.6}.}

\item{Binarization}{Logical indicating if to choose one topic for each
CellTopic. Default is \code{FALSE}.}

\item{meta.cell}{Logical indicating if return MetaTopic, which is a data frame of the
cluster result of CellTopic.}

\item{k}{A integer of how much MetaTopic to choose. If \code{meta.cell = TRUE}, \code{k}
is Requested to given. Default is \code{4}.}

\item{method}{See \code{\link[stats]{hclust}}.}
}
\value{
A list with three data frame and one vector. "CellTopic" is a data
frame which can be add to a Seurat object. The "domain_topic" is a data frame,
row is CellTopic and col is domain. The "celltype_topic" is a data frame, row is
celltype and col is CellTopic. The "Cell_topic" is a vector of which topic be chosen
in each CellTopic. If meta.cell = TRUE, the "MetaTopic" will be given, which
is a data frame of the cluster result of CellTopic.
}
\description{
Through deconvolution and clustering information, the SpaTopic
obtains CellTopics representing the spatial domains.
}
\examples{
options (warn = -1)
result_list <- CellTopic(spot_celltype,
spot_clusters,
cluster = "spatial.cluster",
num_topics = 13,
percent = 0.7,
Binarization = FALSE,
meta.cell = FALSE,
k = NULL)
head(result_list[["CellTopic"]])
print(result_list[["domain_topic"]])
print(result_list[["celltype_topic"]])
print(result_list[["Cell_topic"]])
}
