% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTopic_plot.R
\name{CellTopic_plot}
\alias{CellTopic_plot}
\title{CellTopic Plot}
\usage{
CellTopic_plot(
  st_obj,
  celltype_topic,
  celltopic = colnames(celltype_topic),
  highlight.circle = NULL,
  cols.circle = "black",
  cols.highlight = c("#DE2D26", "grey50"),
  cols.celltype = NULL,
  plot.size = 7,
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{st_obj}{A seurat object with CellTopic meta data.}

\item{celltype_topic}{A data frame, row is celltype and col is CellTopic.}

\item{celltopic}{A character of which CellTopic to plot.}

\item{highlight.circle}{A character of which meta.data to high light. Default is \code{NULL}.}

\item{cols.circle}{A character of circle's color. Default is \code{black}.}

\item{cols.highlight}{A vector with two character of highlight spots' colors.}

\item{cols.celltype}{A vector with character of celltypes' colors. Default is \code{NULL}.}

\item{plot.size}{A numeric of plot's size. Default is \code{7}.}

\item{nrow}{A numeric of rows in the plot grid.}

\item{ncol}{A numeric of cols in the plot grid.}

\item{...}{Arguments passed to other methods.}
}
\value{
A polt with SpatialDimPlot and celltype barpolt.
}
\description{
A drawing function that outputs a CellTopic spatial distribution plot and corresponding celltype barpolt.
}
