% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_expansion.R
\name{Expansion_gene_plot}
\alias{Expansion_gene_plot}
\title{Expansion_gene_plot}
\usage{
Expansion_gene_plot(
  st_obj,
  CellTopic,
  gene,
  meta.col = "CellTopic",
  angle = 90,
  parts = 1,
  plot = TRUE,
  cols = NULL
)
}
\arguments{
\item{st_obj}{A Seurat object.}

\item{CellTopic}{A character of which CellTopic to be chosen.}

\item{gene}{A vector contains one or more genes for display}

\item{meta.col}{A character of which meta.data to be chosen. Default is \code{"CellTopic"}.}

\item{angle}{A numeric indicates the starting angle. Default is \code{90}, which means directly above.}

\item{parts}{A number indicates that the angle needs to be divided into several equal parts.}

\item{plot}{Logical indicating if to display the result plot or just data for
plotting. Default is \code{TRUE}.}

\item{cols}{A vector of colors used for plot.}
}
\value{
If \code{plot = TRUE}, return smooth plots of genes as a function of
buffer distance at different angles. If \code{plot = FALSE}, return the data
for plotting.
}
\description{
Show smooth plots of genes as a function of buffer distance at different angles.
}
