% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTopic_function.R
\name{FindCellTopic}
\alias{FindCellTopic}
\title{FindCellTopic}
\usage{
FindCellTopic(
  dt_topic_data,
  ct_topic_data,
  spot_clusters,
  cluster = NULL,
  percent = NULL,
  Binarization = FALSE
)
}
\arguments{
\item{dt_topic_data}{A data frame, row is topic and col is domain. The result
of function \code{\link[SpaTopic]{MatrixFactorization}}.}

\item{ct_topic_data}{A data frame, row is celltype and col is topic. The
result of function \code{\link[SpaTopic]{MatrixFactorization}}.}

\item{spot_clusters}{A data frame contains clustering information for spots,
row is spots.}

\item{cluster}{A character. Use the first column in spot_clusters as the
column name if \code{NULL}. Or provide your own column names that represent
clustering information in spot_clusters.}

\item{percent}{A numeric from \code{0} to \code{1}. The percent of topics. Default is \code{0.6}.}

\item{Binarization}{Logical indicating if to choose one topic for each
CellTopic. Default is \code{FALSE}.}
}
\value{
A list with three data frame and one vector. "CellTopic" is a data
frame which can be add to a Seurat object. The "domain_topic" is a data frame,
row is CellTopic. and col is domain. The "celltype_topic" is a data frame, row is
celltype and col is CellTopic. The "Cell_topic" is a vector of which topic be chosen
in each CellTopic.
}
\description{
Based on the results of the \code{\link[SpaTopic]{MatrixFactorization}} and
the clustering information, the CellTopics representing each domain are
calculated.
}
