% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frex_topic.R
\name{frex_topic}
\alias{frex_topic}
\title{Calculate FREX ranked words and values for a topic}
\usage{
frex_topic(beta, vocab, topic = 1, top_n = NULL, frex_weight = 0.5)
}
\arguments{
\item{beta}{A numeric matrix of dimension (topics x words) representing the probability
distribution of words within each topic. Each row should sum to 1. Beta must be on
the probability scale (not log scale).}

\item{vocab}{a character vector of vocabulary terms corresponding to the columns of beta.}

\item{topic}{the topic index that we want to calculate, the default is 1.}

\item{top_n}{the number of top words to return, the default is to return all words.}

\item{frex_weight}{the weight between 0 and 1 controlling the balance between
frequency and exclusivity in the FREX metric. Weight closer to 1 is favoring exclusivity
and closer to 0 is favoring frequency, we set the default as 0.5.}
}
\value{
a data frame with ranks, words, and FREX values of the words
}
\description{
Calculate FREX ranked words and values for a topic
}
