% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTopic_function.R
\name{ksTest}
\alias{ksTest}
\title{ksTest}
\usage{
ksTest(spot_celltype, spot_clusters, cluster = NULL)
}
\arguments{
\item{spot_celltype}{A data frame of the result of deconvolution, row
is spot, col is celltype.}

\item{spot_clusters}{A data frame contains clustering information for spots,
row is spots.}

\item{cluster}{A character. Use the first column in spot_clusters as the
column name if \code{NULL}. Or provide your own column names that represent
clustering information in spot_clusters.}
}
\value{
A data frame, row is spot_domain, col is celltype score.
}
\description{
Kolmogorov-Smirnov (KS) test is performed on the deconvolution results and the overall level of
each cluster.
}
