% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_data.R
\docType{data}
\name{lung5}
\alias{lung5}
\title{Example input data for SpaTopic}
\format{
## `lung5`
A data frame with 100149 rows and 4 columns:
}
\source{
<https://nanostring.com/products/cosmx-spatial-molecular-imager/ffpe-dataset/nsclc-ffpe-dataset/>
}
\usage{
lung5
}
\description{
a multiplexed image data on tumor tissue sample from non small cell lung cancer patients
}
\keyword{datasets}
