% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTopic_function.R
\name{MatrixFactorization}
\alias{MatrixFactorization}
\title{MatrixFactorization}
\usage{
MatrixFactorization(domain_cellytpe, num_topics = NULL)
}
\arguments{
\item{domain_cellytpe}{A data frame. The result of function \code{\link[SpaTopic]{ksTest}}.}

\item{num_topics}{A integer or a vector of integer. The number of topics. To
find the best number of topics if \code{NULL}. Or to find the best number of
topics by the given range if a vector of integer. Or provide your own number
if a integer.}
}
\value{
A list with two data frame. The "domain_topic" is a data frame,
row is topic and col is domain. The "celltype_topic" is a data frame, row is
celltype and col is topic.
}
\description{
Perform matrix factorization of the KS test results, see the matrix
factorization method at \code{\link[topicmodels]{LDA}}.
}
