% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTopic_function.R
\name{MetaTopic}
\alias{MetaTopic}
\title{MetaTopic}
\usage{
MetaTopic(ct_topic_data, k, method)
}
\arguments{
\item{ct_topic_data}{A data frame, row is celltype and col is CellTopic.}

\item{k}{A integer of how much MetaTopic to choose.}

\item{method}{See \code{\link[stats]{hclust}}.}
}
\value{
A data frame of the cluster result of CellTopic.
}
\description{
Make a MetaTopic from CellTopic by celltype scores.
}
