% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_topic.R
\name{score_topic}
\alias{score_topic}
\title{Calculate score ranked words and values for a topic}
\usage{
score_topic(beta, vocab, topic = 1, top_n = NULL)
}
\arguments{
\item{beta}{A numeric matrix of dimension (topics x words) representing the probability
distribution of words within each topic. Each row should sum to 1. Beta must be on
the probability scale (not log scale).}

\item{vocab}{a character vector of vocabulary terms corresponding to the columns of beta.}

\item{topic}{the topic index that we want to calculate, the default is 1.}

\item{top_n}{the number of top words to return, the default is to return all words.}
}
\value{
a data frame with ranks, words, and score values of the words
}
\description{
Calculate score ranked words and values for a topic
}
