% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_SpaTopic_input.R
\name{Seurat5obj_to_SpaTopic}
\alias{Seurat5obj_to_SpaTopic}
\title{Convert one Seurat v5 object as the input of SpaTopic}
\usage{
Seurat5obj_to_SpaTopic(object, group.by, image = "image1")
}
\arguments{
\item{object}{Seurat v5 object}

\item{group.by}{\code{character}. The name of the column that contains celltype information}

\item{image}{\code{character}. The name of the image}
}
\value{
Return a data frame as the input of SpaTopic
}
\description{
Prepare SpaTopic input from one Seurat v5 object
}
\examples{

dataset<-Seurat5obj_to_SpaTopic(object = nano.obj, group.by = "predicted.annotation.l1",image = "image1")

}
