% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_expansion.R
\name{Spatial_expansion}
\alias{Spatial_expansion}
\title{Spatial_expansion}
\usage{
Spatial_expansion(
  st_obj,
  CellTopic,
  meta.col = "CellTopic",
  distance = NULL,
  alpha = NULL,
  type = c("Hexagon", "Square"),
  edge.plot = TRUE,
  image.factor = "lowres"
)
}
\arguments{
\item{st_obj}{A Seurat object.}

\item{CellTopic}{A character string specifying the CellTopic to be chosen.}

\item{meta.col}{A character string specifying the meta.data column to be used. Defaults to \code{"CellTopic"}.}

\item{distance}{A numeric value indicating the distance of expansion. Defaults to \code{NULL},
in which case the nearest spot distance will be used.}

\item{alpha}{A numeric value used for alpha-shape calculation. Defaults to \code{NULL},
which uses the nearest spot distance based on data type. See also \code{\link[alphahull]{ashape}} for more details.}

\item{type}{A character string specifying the type of spot arrangement. Defaults to \code{"Hexagon"} for data with hexagonal spot arrangements.
For square spot arrangements, set \code{type = "Square"}.}

\item{edge.plot}{Logical indicating whether to display the regional division plot
for the specified CellTopic. Defaults to \code{TRUE}.}

\item{image.factor}{A character string specifying the resolution scale factor. Defaults to \code{"lowres"}.}
}
\value{
A Seurat object. The expansion of the data is added to the \code{Seurat@meta.data}
and the message is added to \code{Seurat@misc}.
}
\description{
Applies buffering and angle calculations to divide the specified spatial domain within a Seurat object.
}
