% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spot_celltype}
\alias{spot_celltype}
\title{Deconvolution results of single-cell versus spatial transcriptome}
\format{
A data frame with 428 rows and 15 variables
}
\usage{
spot_celltype
}
\description{
A data frame for each cell type in each spot using the CARD deconvolution method
}
\keyword{datasets}
