% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spot_clusters}
\alias{spot_clusters}
\title{Clustering results of the spatial transcriptome}
\format{
A data frame with 428 rows and 5 variables:
\describe{
 \item{cluster.init}{cluster.init}
 \item{col}{col}
 \item{row}{row}
 \item{spatial.cluster}{the clustering result}
 \item{sizeFactor}{sizeFactor}
 ...
}
}
\usage{
spot_clusters
}
\description{
A data frame of clustering result obtained by BayesSpace
}
\keyword{datasets}
