% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_data.R
\docType{data}
\name{lung5}
\alias{lung5}
\title{Example input data for SpaTopic}
\format{
## `lung5`
A data frame with 100149 rows and 4 columns: 
\describe{ 
  \item{image} Image ID
  \item{X} X coordinate of the cell
  \item{Y} Y coordinate of the cell
  \item{type} cell type
}
}
\source{
<https://nanostring.com/products/cosmx-spatial-molecular-imager/ffpe-dataset/nsclc-ffpe-dataset/>
}
\usage{
lung5
}
\description{
multiplexed image data on tumor tissue sample from non small cell lung cancer patient
}
\seealso{
\code{\link{SpaTopic_inference}},\code{\link{Seurat5obj_to_SpaTopic}}
}
\keyword{datasets}
