% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_expansion.R
\name{Spatial_expansion}
\alias{Spatial_expansion}
\title{Spatial_expansion}
\usage{
Spatial_expansion(
  st_obj,
  CellTopic,
  meta.col = "CellTopic",
  distance = NULL,
  alpha = NULL,
  type = c("Hexagon", "Square"),
  edge.plot = TRUE,
  image.factor = "lowres"
)
}
\arguments{
\item{st_obj}{A Seurat object.}

\item{CellTopic}{A character of which CellTopic to be chosen.}

\item{meta.col}{A character of which meta.data to be chosen. Default is \code{"CellTopic"}.}

\item{distance}{A numeric. The distance of expansion. Default is \code{NULL},
use the nearest spot distance.}

\item{alpha}{A numeric used for alpha-shape calculation. Default is \code{NULL},
use nearest spot distance based on data type. See also \code{\link[alphahull]{ashape}}.}

\item{type}{By default \code{type = "Hexagon"} for data with hexagonal spot arrangement.
When the spot arrangement of data is square, set \code{tpye = "Square"}.}

\item{edge.plot}{Logical indicating if to display the regional division plot
of CellTopic. Default is \code{TRUE}.}

\item{image.factor}{A character of the resolution scale factor. Default is \code{"lowres"}.}
}
\value{
A Seurat object. The expansion of the data is added to the \code{Seurat@meta.data}
and the message is added to \code{Seurat@misc}.
}
\description{
Buffers and angles for dividing the specified spatial domain.
}
