\name{SpaTopic-package}
\alias{SpaTopic-package}
\alias{SpaTopic}
\docType{package}
\title{
\packageTitle{SpaTopic}
}
\description{
\packageDescription{SpaTopic}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SpaTopic}
\packageIndices{SpaTopic}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SpaTopic}

Maintainer: \packageMaintainer{SpaTopic}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
