% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaTopic-package.R
\docType{package}
\name{SpaTopic-Package}
\alias{SpaTopic}
\alias{SpaTopic-package}
\alias{SpaTopic-Package}
\title{'SpaTopic' R package}
\description{
The 'SpaTopic' R package is centered around the 'SpaTopic' algorithm to infer the spatial
tissue architectures from multiplexed images.
}
\details{
The package implements a Collapsed 
Gibbs sampling algorithm to infer topics, corresponding to distinct tissue 
microenvironments across multiple tissue images. 
Without obtaining the cell neighborhood info of every single cell, 
'SpaTopic' runs much faster than other KNN-based methods on large-scale images.

The main functions in the 'SpaTopic' package

\itemize{
\item Prepare input   \code{\link{Seurat5obj_to_SpaTopic}}
\item Model Inference \code{\link{SpaTopic_inference}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xiyupeng/SpaTopic}
  \item Report bugs at \url{https://github.com/xiyupeng/SpaTopic/issues}
}

}
\author{
Xiyu Peng \email{pansypeng124@gmail.com}
}
\keyword{package}
